refactor(shadcn-ui): colocate form adapters
