+++ examples/shadcn-ui/src/App.tsx
@@ -25,8 +25,8 @@ import {
 	InputOTP,
 	TeamMemberSelect,
 	memberSchema,
-} from './components/form-controls';
-import { useForm } from './forms';
+	useForm,
+} from './forms';
 
 const schema = coerceFormValue(
 	z.object({
