--- examples/shadcn-ui/src/forms.ts
@@ -1,180 +1,0 @@
-import {
-	configureForms,
-	type FieldMetadata as BaseFieldMetadata,
-	type FormMetadata as BaseFormMetadata,
-	type Fieldset as BaseFieldset,
-	type InferBaseErrorShape,
-	type InferCustomFieldMetadata,
-	type InferCustomFormMetadata,
-} from '@conform-to/react/future';
-import { getConstraints } from '@conform-to/zod/v4/future';
-import type {
-	DatePicker,
-	ComboBox,
-	RadioGroup,
-	Checkbox,
-	Select,
-	Slider,
-	Switch,
-	SingleToggleGroup,
-	MultiToggleGroup,
-	InputOTP,
-	TeamMemberSelect,
-} from './components/form-controls';
-
-const forms = configureForms({
-	getConstraints,
-	shouldValidate: 'onBlur',
-	shouldRevalidate: 'onInput',
-	extendFieldMetadata(metadata) {
-		const getAriaDescribedBy = () =>
-			[metadata.descriptionId, metadata.ariaDescribedBy]
-				.filter(Boolean)
-				.join(' ');
-
-		return {
-			get inputProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<'input'>>;
-			},
-			get textareaProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<'textarea'>>;
-			},
-			get datePickerProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof DatePicker>>;
-			},
-			get comboBoxProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof ComboBox>>;
-			},
-			get radioGroupProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-					'aria-labelledby': `${metadata.id}-label`,
-				} satisfies Partial<React.ComponentProps<typeof RadioGroup>>;
-			},
-			get checkboxProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					value: 'on',
-					defaultChecked: metadata.defaultChecked,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof Checkbox>>;
-			},
-			get selectProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof Select>>;
-			},
-			get sliderProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-					'aria-labelledby': `${metadata.id}-label`,
-				} satisfies Partial<React.ComponentProps<typeof Slider>>;
-			},
-			get switchProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					value: 'on',
-					defaultChecked: metadata.defaultChecked,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof Switch>>;
-			},
-			get singleToggleGroupProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-labelledby': `${metadata.id}-label`,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof SingleToggleGroup>>;
-			},
-			get multiToggleGroupProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultOptions,
-					'aria-labelledby': `${metadata.id}-label`,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof MultiToggleGroup>>;
-			},
-			get inputOTPProps() {
-				return {
-					id: metadata.id,
-					name: metadata.name,
-					defaultValue: metadata.defaultValue,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof InputOTP>>;
-			},
-			get teamMemberSelectProps() {
-				return {
-					name: metadata.name,
-					defaultValue: metadata.defaultPayload,
-					'aria-labelledby': metadata.id,
-					'aria-describedby': getAriaDescribedBy(),
-					'aria-invalid': metadata.ariaInvalid,
-				} satisfies Partial<React.ComponentProps<typeof TeamMemberSelect>>;
-			},
-		};
-	},
-});
-
-type BaseErrorShape = InferBaseErrorShape<typeof forms.config>;
-type CustomFormMetadata = InferCustomFormMetadata<typeof forms.config>;
-type CustomFieldMetadata = InferCustomFieldMetadata<typeof forms.config>;
-
-export type FormMetadata<ErrorShape extends BaseErrorShape = BaseErrorShape> =
-	BaseFormMetadata<ErrorShape, CustomFormMetadata, CustomFieldMetadata>;
-
-export type FieldMetadata<
-	FieldShape,
-	ErrorShape extends BaseErrorShape = BaseErrorShape,
-> = BaseFieldMetadata<FieldShape, ErrorShape, CustomFieldMetadata>;
-
-export type Fieldset<
-	FieldShape,
-	ErrorShape extends BaseErrorShape = BaseErrorShape,
-> = BaseFieldset<FieldShape, ErrorShape, CustomFieldMetadata>;
-
-export const useForm = forms.useForm;
